Added box blur filter

+++ src/filters/utils.rs
@@ -13,15 +13,15 @@ pub fn average_rgb(pixels: &Vec<image::Rgb<u8>>) -> image::Rgb<u8> {
 }
 
 pub fn average_squared_rgb(pixels: &Vec<image::Rgb<u8>>) -> image::Rgb<u8> {
-    let mut rgb_sums = [0, 0, 0];
+    let mut rgb_sums = [0.0, 0.0, 0.0];
     for pixel in pixels {
         pixel
             .0
             .iter()
             .enumerate()
-            .for_each(|(i, &value)| rgb_sums[i] += value.pow(2) as u32);
+            .for_each(|(i, &value)| rgb_sums[i] += (value as f64).powf(2.0));
     }
     let n = pixels.len() as f64;
-    let rgb_means = rgb_sums.map(|v| (v as f64 / n).sqrt() as u8);
+    let rgb_means = rgb_sums.map(|v| (v / n).sqrt() as u8);
     image::Rgb(rgb_means)
 }
